[FrameworkBundle] Fix setting decorated services during tests
