+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/TestServiceContainerTest.php
@@ -16,6 +16,7 @@ use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServic
 use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServiceContainer\PrivateService;
 use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServiceContainer\PublicService;
 use Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\TestBundle\TestServiceContainer\UnusedPrivateService;
+use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 
 class TestServiceContainerTest extends AbstractWebTestCase
 {
@@ -38,6 +39,33 @@ class TestServiceContainerTest extends AbstractWebTestCase
         $this->assertTrue(static::getContainer()->has(PrivateService::class));
         $this->assertTrue(static::getContainer()->has('private_service'));
         $this->assertFalse(static::getContainer()->has(UnusedPrivateService::class));
+    }
+
+    public function testThatPrivateServicesCanBeSetIfTestConfigIsEnabled()
+    {
+        static::bootKernel(['test_case' => 'TestServiceContainer']);
+
+        $container = static::getContainer();
+
+        $service = new \stdClass();
+
+        $container->set('private_service', $service);
+        $this->assertSame($service, $container->get('private_service'));
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The "private_service" service is already initialized, you cannot replace it.');
+        $container->set('private_service', new \stdClass());
+    }
+
+    public function testSetDecoratedService()
+    {
+        static::bootKernel(['test_case' => 'TestServiceContainer']);
+
+        $container = static::getContainer();
+
+        $service = new PrivateService();
+        $container->set('decorated', $service);
+        $this->assertSame($service, $container->get('decorated')->inner);
     }
 
     /**
