+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -21,7 +21,7 @@
         "ext-xml": "*",
         "symfony/cache": "^5.4|^6.0",
         "symfony/config": "^6.1",
-        "symfony/dependency-injection": "^6.3",
+        "symfony/dependency-injection": "^6.3.1",
         "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/error-handler": "^6.1",
         "symfony/event-dispatcher": "^5.4|^6.0",
