+++ src/Symfony/Component/DependencyInjection/Compiler/DecoratorServicePass.php
@@ -42,6 +42,7 @@ class DecoratorServicePass extends AbstractRecursivePass
             $definitions->insert([$id, $definition], [$decorated[2], --$order]);
         }
         $decoratingDefinitions = [];
+        $decoratedIds = [];
 
         $tagsToKeep = $container->hasParameter('container.behavior_describing_tags')
             ? $container->getParameter('container.behavior_describing_tags')
@@ -58,6 +59,7 @@ class DecoratorServicePass extends AbstractRecursivePass
                 $renamedId = $id.'.inner';
             }
 
+            $decoratedIds[$inner] ??= $renamedId;
             $this->currentId = $renamedId;
             $this->processValue($definition);
 
@@ -114,7 +116,7 @@ class DecoratorServicePass extends AbstractRecursivePass
         }
 
         foreach ($decoratingDefinitions as $inner => $definition) {
-            $definition->addTag('container.decorator', ['id' => $inner]);
+            $definition->addTag('container.decorator', ['id' => $inner, 'inner' => $decoratedIds[$inner]]);
         }
     }
